+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.2] - 2026-09-20 @ 12:30
+
 ### Changed
 
 - Mark seed changes in the report server's Experiment Highscores plot with purple points, a hover label, and separate legend entries for the blue score line, orange accepted configurations, and purple seed changes.